Instrument consume call in Azure Service Bus
