--- src/Tingle.EventBus.Transports.Amazon.Sqs/AttributeNames.cs
@@ -1,10 +1,0 @@
-﻿namespace Tingle.EventBus.Transports.Amazon.Sqs
-{
-    internal static class AttributeNames
-    {
-        public const string ContentType = "Content-Type";
-        public const string SequenceNumber = "SequenceNumber";
-        public const string CorrelationId = "CorrelationId";
-        public const string ActivityId = "EventBus.ActivityId";
-    }
-}
