+++ src/Tingle.EventBus/AttributeNames.cs
@@ -1,0 +1,17 @@
+﻿namespace Tingle.EventBus
+{
+    /// <summary>
+    /// Attribute and property names added alongside a message/event if the transport supports
+    /// </summary>
+    internal static class AttributeNames
+    {
+        public const string Id = "Id";
+        public const string ContentType = "Content-Type";
+        public const string SequenceNumber = "SequenceNumber";
+        public const string CorrelationId = "CorrelationId";
+
+        public const string ActivityId = "EventBus.ActivityId";
+        public const string EventType = "EventBus.EventType";
+        public const string EventName = "EventBus.EventName";
+    }
+}
